create prod_community dart file
