+++ app/lib/config/consts.dart
@@ -1,6 +1,7 @@
 import 'dart:io';
 
 import 'package:encointer_wallet/config/node.dart';
+import 'package:encointer_wallet/config/prod_community.dart';
 import 'package:encointer_wallet/service/log/log_service.dart';
 import 'package:encointer_wallet/store/settings.dart';
 
@@ -142,33 +143,3 @@ String replaceLocalePlaceholder(String link, String locale) {
       return link.replaceAll(localePlaceHolder, 'en');
   }
 }
-
-enum ProdCommunity {
-  leo(notificationSound: _leuZurichSound, webSiteLink: _leuZurichLink),
-  gbd(notificationSound: _greenbaySound, webSiteLink: _greenbayLink);
-
-  const ProdCommunity({
-    required this.webSiteLink,
-    required this.notificationSound,
-  });
-
-  final String webSiteLink;
-  final String notificationSound;
-
-  static ProdCommunity getCommunityByCid(String? cid) {
-    switch (cid) {
-      case 'dpcmj33LUs9':
-      case 'dpcm5272THU':
-        return gbd;
-      case 'u0qj944rhWE':
-        return leo;
-      default:
-        return leo;
-    }
-  }
-}
-
-const _leuZurichSound = 'lions_growl';
-const _greenbaySound = 'gbd_chime';
-const _leuZurichLink = 'https://leu.zuerich/$localePlaceHolder';
-const _greenbayLink = 'http://greenbaydollar.com/';
